+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to tlgram will be documented in this file.
 
+## [0.6] - 2026-01-16
+
+### Added
+- Self-update support (closes #5)
+  - `--check-update` flag to check if a new version is available
+  - `--update` flag to update tlgram to the latest version
+
 ## [0.5] - 2026-01-16
 
 ### Added